ref: minor code changes
